refactor: move hints to game feature
