+++ src/features/game/model/HintSettings.ts
