+++ src/features/game/ui/card/WordCard.module.css
