+++ src/features/game/ui/rows/Row.ts
@@ -1,11 +1,11 @@
-import { Component, div } from "../../../shared/ui";
-import { debounceListener } from "../../../shared/helpers";
-import { Observer, Publisher } from "../../../entities/state";
+import { Component, div } from "../../../../shared/ui";
+import { debounceListener } from "../../../../shared/helpers";
+import { Observer, Publisher } from "../../../../entities/state";
 
 import WordCard, { type Word } from "../card/WordCard";
 
-import { HintSettings } from "../../hints";
-import RoundState from "../model/RoundState";
+import HintSettings from "../../model/HintSettings";
+import RoundState from "../../model/RoundState";
 
 import styles from "./Row.module.css";
 import { RowType } from "./RowType";
