--- src/features/hints/index.ts
@@ -1,6 +1,0 @@
-export {
-  default as HintSettings,
-  type HintSettingsData,
-} from "./model/HintSettings";
-
-export { default as HintControls } from "./ui/HintControls";
